pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel", "setuptools-git-versioning>=2.0,<3"]
build-backend = "setuptools.build_meta"

[project]
name = "uk-power-flow-to-solar-capacity"
dynamic = ["version"]
description = "Temporal Fusion Transformer for UK solar power forecasting using UKPN and ERA5 data."
readme = {file = "README.md", content-type = "text/markdown"}
requires-python = ">=3.12.0"
license = {text = "MIT License"}

authors = [
    { name="Felix Peretz", email="felix@openclimatefix.org" }
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
]

dependencies = [
    "pandas >= 2.0.0",
    "numpy >= 1.23.2",
    "dask[dataframe] >= 2023.1.0",
    "pyarrow >= 10.0.0",
    "torch >= 2.0.0",
    "pytorch-lightning >= 2.0.0",
    "lightning >= 2.0.0",
    "pytorch-forecasting >= 1.0.0",
    "scikit-learn >= 1.3.0",
    "optuna >= 3.0.0",
    "ray[tune] >= 2.0.0",
    "xarray >= 2023.1.0",
    "netCDF4 >= 1.6.0",
    "pyproj >= 3.4.0",
    "ephem >= 4.1.0",
    "matplotlib >= 3.6.0",
    "seaborn >= 0.12.0",
    "plotly >= 5.0.0",
    "pyyaml >= 6.0.0",
    "loguru >= 0.7.3",
    "hydra-core >= 1.3.0",
    "holidays >= 0.34.0",
    "joblib >= 1.3.0",
    "tqdm >= 4.65.0",
    "pvlib >= 0.10.0",
    "astral >= 3.2.0",
    "pydantic >= 2.0.0",
    "pydantic-settings >= 2.0.0",
    "geopandas >= 0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest >= 7.0.0",
    "pytest-cov >= 4.0.0",
    "pytest-xdist >= 3.0.0",
    "unittest-xml-reporting >= 3.2.0",
    "mypy >= 1.5.0",
    "ruff >= 0.9.2",
    "black >= 23.0.0",
    "sphinx >= 7.0.0",
    "sphinx-rtd-theme >= 1.3.0",
    "myst-parser >= 2.0.0",
    "jupyter >= 1.0.0",
    "ipykernel >= 6.25.0",
    "ipywidgets >= 8.1.0",
    "python-lsp-server",
    "pylsp-mypy",
    "python-lsp-ruff",
]

gpu = [
    "torch[cuda]",
    "torchvision",
    "torchaudio",
]

cloud = [
    "google-cloud-storage >= 2.10.0",
    "google-cloud-bigquery >= 3.11.0",
    "gcsfs >= 2023.6.0",
    "dask[distributed] >= 2023.1.0",
    "fsspec[gcs] >= 2023.6.0",
]

all = [
    "uk-power-flow-to-solar-capacity[dev,gpu,cloud]"
]

[tool.setuptools]
include-package-data = true
license-files = ["LICENSE"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools-git-versioning]
enabled = true

[tool.mypy]
python_version = "3.12"
strict = true
warn_unreachable = true
warn_return_any = false
disallow_untyped_defs = true
check_untyped_defs = true
show_error_codes = true
plugins = [
    "numpy.typing.mypy_plugin",
]

[[tool.mypy.overrides]]
module = [
    "pytorch_lightning.*",
    "lightning.*",
    "pytorch_forecasting.*",
    "optuna.*",
    "ray.*",
    "pvlib.*",
    "ephem.*",
    "astral.*",
    "pyproj.*",
    "holidays.*",
    "torch.*",
    "torchvision.*",
    "torchaudio.*",
    "sklearn.*",
    "xarray.*",
    "netCDF4.*",
    "matplotlib.*",
    "seaborn.*",
    "plotly.*",
    "dask.*",
    "omegaconf.*",
    "hydra.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py312"
preview = true

[tool.ruff.lint]
select = [
    "F", "E", "W", "I", "N", "UP", "YTT", "ANN", "S", "BLE", "B", 
    "A", "C4", "T10", "ISC", "ICN", "PT", "SIM", "TID", "ARG", "DTZ", 
    "PD", "RUF"
]
ignore = [
    "ANN101", "ANN102", "D", "COM812", "ISC001", "T201", "T203"
]

[tool.ruff.lint.isort]
known-first-party = ["src"]
sections = { "pytorch" = ["torch", "pytorch_lightning", "lightning", "pytorch_forecasting"] }
section-order = ["future", "standard-library", "pytorch", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "ANN"]
"notebooks/*" = ["ALL"]
"experiments/*" = ["T20", "F841", "F401"]
"src/tft/data.py" = ["PD002", "PD010"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "--strict-markers",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__init__.py"]
